experiment/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
celery==5.3.4
redis==5.0.1
python-telegram-bot==20.7
google-api-python-client==2.108.0
google-auth==2.25.2
google-auth-httplib2==0.1.1
httpx==0.25.2
python-dotenv==1.0.0
phonenumbers==8.13.27
bleach==6.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
aiosqlite==0.19.0
httpx==0.25.2

